Dockerfile.dev: use configurable user id
Before commands inside the container were run by root. This will set the
user of container generated files to root which means they have wrong
permissions. This change allows to define a user id when building the
dev container which will later be used to run applications in the
container.

Signed-off-by: Stefan Eichenberger <stefan.eichenberger@toradex.com>